feat: add domain agents for review workflow including summarization and GitHub commenting
