+++ crates/forge_review/src/domain/bug_reporter.rs
@@ -1,21 +1,19 @@
 use std::sync::Arc;
 
-use crate::domain::PullRequest;
+use super::{PullRequest, SummaryAgent};
 
-use super::SummaryAgent;
-
-pub struct ArchitectureAgent {
+pub struct BugReporterAgent {
     review: Arc<PullRequest>,
 }
 
-impl ArchitectureAgent {
+impl BugReporterAgent {
     pub fn new(review: Arc<PullRequest>) -> Self {
         Self { review }
     }
 }
 
 #[async_trait::async_trait]
-impl SummaryAgent for ArchitectureAgent {
+impl SummaryAgent for BugReporterAgent {
     async fn summarize(&self) -> anyhow::Result<String> {
         todo!()
     }
