+++ crates/forge_review/src/lib.rs
@@ -1,6 +1,5 @@
-mod agents;
 mod domain;
 mod github;
 mod infra;
 
-pub use agents::ReviewWorkflow;
+pub use domain::ReviewWorkflow;
